fix Device.IP.Interface.i.Stats.BytesSent and BytesReceived give wrong values after high traffic ( OFW-5109 ) - part 2

+++ source/TR-181/middle_layer_src/cosa_ip_dml.c
@@ -6042,14 +6042,14 @@ Stats5_GetParamUlongValue
     if (strcmp(ParamName, "PacketsSent") == 0)
     {
         /* collect value */
-        *puLong = pIPStats->PacketsSent;
+        *puLong = (ULONG) pIPStats->PacketsSent;
         goto SUCCESS;
     }
 
     if (strcmp(ParamName, "PacketsReceived") == 0)
     {
         /* collect value */
-        *puLong = pIPStats->PacketsReceived;
+        *puLong = (ULONG) pIPStats->PacketsReceived;
         goto SUCCESS;
     }
